# perf(cli): enable node 22 compile cache