--- a/lib/jiti-cli.mjs
+++ b/lib/jiti-cli.mjs
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 import { resolve } from "node:path";
-import { createJiti } from "./jiti.cjs";
+import nodeModule from "node:module";
 
 const script = process.argv.splice(2, 1)[0];
 
@@ -10,8 +10,26 @@ if (!script) {
   process.exit(1);
 }
 
+// https://nodejs.org/api/module.html#moduleenablecompilecachecachedir
+// https://github.com/nodejs/node/pull/54501
+if (nodeModule.enableCompileCache && !process.env.NODE_DISABLE_COMPILE_CACHE) {
+  try {
+    nodeModule.enableCompileCache();
+  } catch {
+    // Ignore errors
+  }
+}
+
 const pwd = process.cwd();
+
+const { createJiti } = nodeModule.createRequire(import.meta.url)("./jiti.cjs");
+// const { createJiti } = await import("./jiti.cjs");
+
 const jiti = createJiti(pwd);
+
 const resolved = (process.argv[1] = jiti.resolve(resolve(pwd, script)));
 
-jiti.import(resolved).catch(console.error);
+await jiti.import(resolved).catch((error) => {
+  console.error(error);
+  process.exit(1);
+});