--- a/lib/jiti-cli.mjs
+++ b/lib/jiti-cli.mjs
@@ -22,8 +22,7 @@ if (nodeModule.enableCompileCache && !process.env.NODE_DISABLE_COMPILE_CACHE) {
 
 const pwd = process.cwd();
 
-const { createJiti } = nodeModule.createRequire(import.meta.url)("./jiti.cjs");
-// const { createJiti } = await import("./jiti.cjs");
+const { createJiti } = await import("./jiti.cjs");
 
 const jiti = createJiti(pwd);