… test/__snapshots__/fixtures.test.ts.snap