Leverage new `module.enableCompileCache` for Node.js 22+ (more info https://github.com/nodejs/node/pull/54501)

Speedup is about %17 in Node.js v22.9.0 (Intel(R) Core(TM) i7-10700K CPU @ 3.80GHz perf mode)

```
hyperfine --warmup 3 'node ./lib/jiti-cli.mjs test/fixtures/typescript' 'NODE_DISABLE_COMPILE_CACHE=1 node ./lib/jiti-cli.mjs test/fixtures/typescript'
```

```
Benchmark 1: node ./lib/jiti-cli.mjs test/fixtures/typescript
  Time (mean ± σ):     130.3 ms ±   0.8 ms    [User: 131.3 ms, System: 18.0 ms]
  Range (min … max):   128.8 ms … 131.3 ms    22 runs
 
Benchmark 2: NODE_DISABLE_COMPILE_CACHE=1 node ./lib/jiti-cli.mjs test/fixtures/typescript
  Time (mean ± σ):     152.0 ms ±   1.1 ms    [User: 149.2 ms, System: 22.0 ms]
  Range (min … max):   150.1 ms … 154.0 ms    19 runs
 
Summary
  'node ./lib/jiti-cli.mjs test/fixtures/typescript' ran
    1.17 ± 0.01 times faster than 'NODE_DISABLE_COMPILE_CACHE=1 node ./lib/jiti-cli.mjs test/fixtures/typescript'
```